Small rendering cleanup & bugfix
1. Cleanup

The reason for the difference between get-rendered-entries and
build-page was not obvious, so put them near each other and make that
clear. They should be refactored to avoid duplication soon, though they
aren't completely a silly overlap.

2. Bugfix

Deleting the "raw-body" value meant that the cached entries collection
was being mutated, which was bad. This caused the private reference or
public html - whichever was generated second - to not have content at
all. Glad I added the testing!

-POLM