Remove eval-ls functionality
Originally I thought it would be nice to be able to embed arbitrary
computation, but this wasn't a very good way to do it. In practice the
only place I used it was on the top page to show just the first five
items from the archive. I replaced that with a builtin command
("recent").

Embedding code is still something I'd like to do, but it needs a
different approach. -POLM