Move caching into main deltos command
This was a tiny bash script but no real reason for that.

Also renamed all-to-json to dump-json to be more in line with dump-tsv
etc.

Want to move search into the main script as well, but tty/curses issues
make that tricky using percol - maybe used blessed or something. -POLM