A number of search and tag-related fixes
- dumping search json for only public articles works now

Previously the json cache was assumed to be dumb and only used
privately.

- removed configurable published/hidden tags

I didn't think this through when I first added it, but there's no
obvious reason to have configurable tags for publishing or hiding
articles. The handling of this is still too distributed but at least the
config isn't involved any more. -POLM