Make standard search use philtre
For many cases, this doesn't change the items that hit a query. However,
filtering (with negative conditions) is much easier, as is building more
complex queries.

One downside of using philtre is that matches are not highlighted. That
needs more work in searchy. For now this is fine though. -POLM