Add quick dump of all articles, use it for tsv
The default function for loading all entries is slow for several
reasons.

- it resolves child/parent relations
- it sorts entries by date (which doesn't match the filesystem)

This new function just dumps files as they are read from disk, which
means output begins almost immediately. Currently it's just used in the
tsv function, which is now the input to search, but its use should be
evaluated in other places where the parent/child relations aren't
necessary. -POLM