Search & Thread Improvements
Search is faster.

1. Search now has its own file to avoid circular requirements.
2. TSV dump is generated from sqlite
3. TSV dump doesn't spawn a deltos subprocess

In testing this brings time to render down by ~50% to a bit over half a
second with 1500 entries.

The thread interface has been simplified. Deltos just provides a sorted
list of post ids now, and whatever tool want to use the thread has to
find the id it wants. This may later change to produce TSV output for
the thread.

-POLM